Move sniffer callback traits into detail/type_traits.h
